+++ src/into/into_result.rs
@@ -8,6 +8,10 @@ pub trait IntoResult {
     /// # Type Parameters
     /// - `E`: The error type to use in the [Result].
     ///
+    /// # Errors
+    ///
+    /// Never returns an error.
+    ///
     /// # Example
     /// ```rust
     /// use fluent_result::IntoResult;
@@ -31,6 +35,10 @@ pub trait IntoResult {
     /// # Type Parameters
     /// - `T`: The success type to use in the [Result].
     ///
+    /// # Errors
+    ///
+    /// Returns an [Err] variant of a [Result] with the given error value.
+    ///
     /// # Example
     /// ```rust
     /// use fluent_result::IntoResult;
@@ -53,20 +61,3 @@ pub trait IntoResult {
 /// # Type Parameters
 /// - `T`: The success type to use in the `Result`.
 impl<T> IntoResult for T {}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn into_ok() {
-        let ok: Result<u32, ()> = 42.into_ok();
-        assert_eq!(ok, Ok(42));
-    }
-
-    #[test]
-    fn into_err() {
-        let err = "error".into_err::<()>();
-        assert_eq!(err, Err("error"));
-    }
-}