+++ src/map/mod.rs
@@ -1,5 +1,0 @@
-pub mod option_map_to;
-pub mod result_map_to;
-
-pub use option_map_to::OptionMapTo;
-pub use result_map_to::ResultMapTo;