+++ src/nested/mod.rs
@@ -1,0 +1,5 @@
+mod box_err;
+mod flatten_err;
+
+pub use box_err::BoxErr;
+pub use flatten_err::FlattenErr;